Refactor package structure: move ConfigurationCommand to commands package
